Convert to TS
For consistency
