+++ e2e/browser/testApp/package.json
@@ -19,19 +19,13 @@
     "@inrupt/solid-client": "^1.20.1",
     "@inrupt/solid-client-access-grants": "file:../../../",
     "@inrupt/solid-client-authn-browser": "^1.12.2",
-    "next": "^12.1.0",
-    "react": "^16.14.0",
-    "react-dom": "^16.14.0"
+    "next": "^13.0.4",
+    "react": "^18.2.0"
   },
   "devDependencies": {
-    "@inrupt/eslint-config-lib": "^0.4.1",
+    "@types/node": "^18.11.9",
+    "@types/react": "^18.0.25",
     "eslint-config-next": "^12.1.5",
-    "typescript": "^4.2.3"
-  },
-  "overrides": {
-    "next@^12.1.0": {
-      "react": "^16.14.0",
-      "react-dom": "^16.14.0"
-    }
+    "typescript": "^4.9.3"
   }
 }
